Bug 1492365 - Prevent arithmetic overflow. r=gerald
Should mStreamLength be > 2^32, we could have overflowed leading to false positive test.

Differential Revision: https://phabricator.services.mozilla.com/D6235

UltraBlame original commit: e52cb4ceec89c50c113a2787d018ea2783faf70c

+++ dom/media/MediaCache.cpp
@@ -2632,7 +2632,7 @@ MediaCacheStream::ReadBlockFromCache(AutoLock& aLock,
   }
 
   if (mStreamLength >= 0 &&
-      aBuffer.Length() > uint32_t(mStreamLength - aOffset)) {
+      int64_t(aBuffer.Length()) > mStreamLength - aOffset) {
     
     aBuffer = aBuffer.First(mStreamLength - aOffset);
   }
